% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_calcBedpe.R
\name{bb_calcBedpe}
\alias{bb_calcBedpe}
\title{Calculates all potential paired interactions from a BED file}
\usage{
bb_calcBedpe(
  bed,
  res = 10000,
  x = 100,
  decreasing = TRUE,
  interactions = "interchromosomal"
)
}
\arguments{
\item{bed}{data.frame in narrowpeak format}

\item{res}{bin resolution for output BEDPE}

\item{x}{integer top x numer of strongest or weakest peaks}

\item{decreasing}{logical order by decreasing}

\item{interactions}{string filter for interchromosomal or intrachromosomal interactions}
}
\value{
Returns a data.table of BEDPE interactions
}
\description{
Calculates all potential paired interactions from a BED file
}
