% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_copyGroup.R
\name{bb_copyGroup}
\alias{bb_copyGroup}
\title{copies group into another group object so same group can be placed and/or rescaled}
\usage{
bb_copyGroup(group)
}
\arguments{
\item{group}{a "bb_group" object, as outputted from bb_groupPlots}
}
\value{
Function will return an updated "bb_group" object
}
\description{
copies group into another group object so same group can be placed and/or rescaled
}
