% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_filterBedpe.R
\name{bb_filterBedpe}
\alias{bb_filterBedpe}
\title{Filters out short BEDPE interactions}
\usage{
bb_filterBedpe(bedpe, res, buffer)
}
\arguments{
\item{bedpe}{data.frame or data.table in BEDPE format}

\item{res}{integer resolution of bedpe bins}

\item{buffer}{integer number of res-length bins from the center pixel}
}
\value{
Returns a filtered BEDPE
}
\description{
Interactions from a BEDPE file that would result in APA plots intersecting the diagonal
(given a resolution and buffer size) are removed from the BEDPE file.
}
