% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_plotApa.R
\name{bb_plotApa}
\alias{bb_plotApa}
\title{Plots Aggregate Peak Analysis Matrix}
\usage{
bb_plotApa(
  params = NULL,
  apa,
  x = NULL,
  y = NULL,
  width = NULL,
  height = NULL,
  just = c("left", "top"),
  default.units = "inches",
  draw = TRUE,
  palette = colorRampPalette(c("white", "dark red")),
  n = NULL,
  zrange = NULL
)
}
\arguments{
\item{params}{optional "bb_params" object containing relevant function parameters}

\item{apa}{matrix, list of matricies, or 3 column data.frame of APA results}

\item{x}{numeric or unit object specifying x-location of plot}

\item{y}{numeric or unit object specifying y-location of plot}

\item{width}{numeric or unit object specifying width of plot}

\item{height}{numeric or unit object specifying height of plot}

\item{just}{string or numeric vector specifying the justification of the viewport
relative to its (x, y) location}

\item{default.units}{string indicating the default units to use if x, y, width, or
height are only given as numeric vectors}

\item{draw}{logical value indicating whether graphics output should be produced}

\item{palette}{colorRampPalette function to use to map values to colors}

\item{n}{numeric by which to divide aggregate matrix}

\item{zrange}{vector of length 2; max and min values to set color scale}
}
\value{
Function will draw a APA matrix and return an object of class "bb_apa"
}
\description{
Plots Aggregate Peak Analysis Matrix
}
