% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_removeGroup.R
\name{bb_removeGroup}
\alias{bb_removeGroup}
\title{removes a "bb_group" of BentoBox plots}
\usage{
bb_removeGroup(group)
}
\arguments{
\item{group}{a "bb_group" object, as outputted from bb_groupPlots}
}
\description{
removes a "bb_group" of BentoBox plots
}
