% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_ungroupPlots.R
\name{bb_ungroupPlots}
\alias{bb_ungroupPlots}
\title{ungroups a "bb_group" object}
\usage{
bb_ungroupPlots(group)
}
\arguments{
\item{group}{a "bb_group" object, as outputted from bb_groupPlots}
}
\value{
Function will return a list of plot objects that were in the "bb_group"
}
\description{
ungroups a "bb_group" object
}
